# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Video Streaming Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface LoginRequest {
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface RegisterRequest {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  createdAt: string;
}

export interface LoginResponse {
  token: string;
  user: User;
}

export type VideoStatus = (typeof VideoStatus)[keyof typeof VideoStatus];

export const VideoStatus = {
  processing: "processing",
  ready: "ready",
  failed: "failed",
} as const;

export interface Video {
  id: number;
  title: string;
  thumbnailUrl?: string | null;
  /** Duration in seconds */
  duration?: number | null;
  status: VideoStatus;
  uploadedAt: string;
  userId: number;
}

export interface VideoManifest {
  videoId: number;
  manifestUrl: string;
  title: string;
}

export interface VideoUploadRequest {
  title: string;
  file: Blob;
}