# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Video Streaming Platform API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Authenticate user with email and password
 * @summary User login
 */
export const loginBodyPasswordMin = 6;

export const LoginBody = zod.object({
  email: zod.string().email(),
  password: zod.string().min(loginBodyPasswordMin),
});

export const LoginResponse = zod.object({
  token: zod.string(),
  user: zod.object({
    id: zod.number(),
    email: zod.string(),
    name: zod.string(),
    createdAt: zod.date(),
  }),
});

/**
 * Register a new user account
 * @summary User registration
 */
export const registerBodyPasswordMin = 6;

export const RegisterBody = zod.object({
  email: zod.string().email(),
  password: zod.string().min(registerBodyPasswordMin),
  name: zod.string(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  email: zod.string(),
  name: zod.string(),
  createdAt: zod.date(),
});

/**
 * @summary List all videos
 */
export const ListVideosResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  thumbnailUrl: zod.string().nullish(),
  duration: zod.number().nullish().describe("Duration in seconds"),
  status: zod.enum(["processing", "ready", "failed"]),
  uploadedAt: zod.date(),
  userId: zod.number(),
});
export const ListVideosResponse = zod.array(ListVideosResponseItem);

/**
 * @summary Upload a video
 */
export const UploadVideoBody = zod.object({
  title: zod.string(),
  file: zod.instanceof(File),
});

/**
 * @summary Get video by ID
 */
export const GetVideoParams = zod.object({
  id: zod.coerce.string(),
});

export const GetVideoResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  thumbnailUrl: zod.string().nullish(),
  duration: zod.number().nullish().describe("Duration in seconds"),
  status: zod.enum(["processing", "ready", "failed"]),
  uploadedAt: zod.date(),
  userId: zod.number(),
});

/**
 * @summary Get video HLS manifest URL
 */
export const GetVideoManifestParams = zod.object({
  id: zod.coerce.string(),
});

export const GetVideoManifestResponse = zod.object({
  videoId: zod.number(),
  manifestUrl: zod.string(),
  title: zod.string(),
});