# lib/api-zod/src/generated/types/loginResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Video Streaming Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from "./user";

export interface LoginResponse {
  token: string;
  user: User;
}