# lib/api-zod/src/generated/types/video.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Video Streaming Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { VideoStatus } from "./videoStatus";

export interface Video {
  id: number;
  title: string;
  thumbnailUrl?: string | null;
  /** Duration in seconds */
  duration?: number | null;
  status: VideoStatus;
  uploadedAt: Date;
  userId: number;
}