# lib/api-zod/src/generated/types/videoUploadRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Video Streaming Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface VideoUploadRequest {
  title: string;
  file: Blob;
}